Add WebSocket support and refactor MIDI handling
Introduces WebSocket models and service for OBS integration, updates HomeViewModel and Home.razor to reflect WebSocket connection status, and refactors MIDI handling to use byte types for pad positions. Removes unused code and streamlines device status updates.

+++ PadConnect/Components/Models/PanelModel.cs
@@ -10,8 +10,8 @@ namespace PadConnect.Components.Models
 {
     public class PanelModel : INotifyPropertyChanged
     {
-        public int X { get; set; }
-        public int Y { get; set; }
+        public byte X { get; set; }
+        public byte Y { get; set; }
         private bool _isActive;
         public bool IsActive
         {
